Transaction v2

+++ src/database/migrations/20230718153631-create_transaction_toppings_table.js
@@ -18,7 +18,7 @@ module.exports = {
         type: Sequelize.INTEGER,
       },
       total_item: {
-        type: Sequelize.DECIMAL(10, 2),
+        type: Sequelize.INTEGER,
         allowNull: true,
         defaultValue: 0, 
       },
